feat: move rules and small updates to loading
